requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==2.3.7
email-validator==2.1.0
Pillow>=9.0.0,<11.0.0
pandas>=1.5.0
openpyxl>=3.0.0